refactor: 패키지 분리
- view, domain 패키지 분리
