+++ src/main/java/lotto/domain/LottoMessage.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.stream.Collectors;
