+++ src/test/java/lotto/LottoNumberTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.GenerateLottoNumber;
+import lotto.domain.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
